old structure
Can't load new structure in scene builder
