+++ src/main/java/org/example/hakmana/DashboardCardController.java
@@ -1,4 +1,4 @@
-package org.example.hakmana.componentControllers;
+package org.example.hakmana;
 
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -10,29 +10,21 @@ import javafx.scene.control.ScrollPane;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 
-import java.io.File;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
 public class DashboardCardController extends ScrollPane implements Initializable {
+
+
+
     @FXML
     private Label topLabel;
     private String text;
 
-    //methods for component initialize and constructor for component load----------------------------
-    @Override
-    public void initialize(URL url, ResourceBundle resourceBundle) {}
     public DashboardCardController() {
         super();
-        String filePath = "src/main/resources/org/example/hakmana/Component/dashboardCard.fxml";
-        FXMLLoader fxmlDashboardCardLoader = null;
-        try {
-            fxmlDashboardCardLoader = new FXMLLoader(new File(filePath).toURI().toURL());
-        } catch (MalformedURLException e) {
-            throw new RuntimeException(e);
-        }
+        FXMLLoader fxmlDashboardCardLoader=new FXMLLoader(getClass().getResource("Component/dashboardCard.fxml"));
 
         fxmlDashboardCardLoader.setController(this);
         fxmlDashboardCardLoader.setRoot(this);
@@ -48,7 +40,7 @@ public class DashboardCardController extends ScrollPane implements Initializable
     }
 
 
-    //getters and setters for label set----------------------------------------------------
+
     public void setText(String text) {
         this.text = text;
         topLabel.setText(this.text);
@@ -57,4 +49,9 @@ public class DashboardCardController extends ScrollPane implements Initializable
         return topLabel.getText();
     }
 
+
+    @Override
+    public void initialize(URL url, ResourceBundle resourceBundle) {
+
+    }
 }
