+++ src/main/java/org/example/hakmana/DesktopFormController.java
@@ -1,4 +1,4 @@
-package org.example.hakmana.view;
+package org.example.hakmana;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -14,10 +14,9 @@ import org.example.hakmana.model.DatabaseConnection;
 import org.example.hakmana.model.Desktop;
 import org.example.hakmana.model.User;
 
-import java.io.File;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
+import java.sql.SQLOutput;
 import java.util.ResourceBundle;
 
 public class DesktopFormController implements Initializable {
@@ -98,17 +97,10 @@ public class DesktopFormController implements Initializable {
 
     public static User user;
     public void addUser(ActionEvent event) throws IOException {
-        ((Node) event.getSource()).setDisable(UserAssignViewController.isAssignUserButtonClicked);
+        ((Node) event.getSource()).setDisable(UserController.isAssignUserButtonClicked);
 
-        if(!UserAssignViewController.isAssignUserButtonClicked){
-            String filePath = "src/main/resources/org/example/hakmana/Component/Header.fxml";
-            FXMLLoader fxmlHeaderLoader = null;
-            try {
-                fxmlHeaderLoader = new FXMLLoader(new File(filePath).toURI().toURL());
-            } catch (MalformedURLException e) {
-                throw new RuntimeException(e);
-            }
-            Parent root = FXMLLoader.load(getClass().getResource("Scene/DialogBoxView/UserAssignDialog.fxml"));
+        if(!UserController.isAssignUserButtonClicked){
+            Parent root = FXMLLoader.load(getClass().getResource("Scene/user.fxml"));
             Stage stage=new Stage();
             stage.initModality(Modality.APPLICATION_MODAL);
             stage.setScene(new Scene(root));
@@ -116,6 +108,8 @@ public class DesktopFormController implements Initializable {
         }
 
     }
+
+
 
     public void desktop(ActionEvent event) throws IOException {
         Parent root = FXMLLoader.load(getClass().getResource("Scene/desktop.fxml"));
@@ -125,6 +119,7 @@ public class DesktopFormController implements Initializable {
         stage.show();
 
     }
+
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
@@ -212,15 +207,16 @@ public class DesktopFormController implements Initializable {
         NetworkCardChoiseBox.getSelectionModel().clearSelection();
         micRegNumTextField.clear();
 
-        UserAssignViewController.isAssignUserButtonClicked=false;
+        UserController.isAssignUserButtonClicked=false;
         addUserButton.setDisable(false);
 
 
     }
-    public void cancelButtonOnAction(ActionEvent event) {
+ public void cancelButtonOnAction(ActionEvent event) {
      Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
      stage.close();
 
-    }
+ }
+
 
 }
