+++ src/main/java/org/example/hakmana/DeviceInfoCardController.java
@@ -1,4 +1,4 @@
-package org.example.hakmana.componentControllers;
+package org.example.hakmana;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -14,9 +14,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
 
-import java.io.File;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
@@ -30,13 +28,7 @@ public class DeviceInfoCardController extends AnchorPane implements Initializabl
      }
      public DeviceInfoCardController() {
           super();
-          String filePath = "src/main/resources/org/example/hakmana/Component/DeviceInfoCard.fxml";
-          FXMLLoader fxmlFooterLoader = null;
-          try {
-               fxmlFooterLoader = new FXMLLoader(new File(filePath).toURI().toURL());
-          } catch (MalformedURLException e) {
-               throw new RuntimeException(e);
-          }
+          FXMLLoader fxmlFooterLoader = new FXMLLoader(getClass().getResource("Component/DeviceInfoCard.fxml"));
           fxmlFooterLoader.setController(this);
           fxmlFooterLoader.setRoot(this);
 
@@ -116,17 +108,13 @@ public class DeviceInfoCardController extends AnchorPane implements Initializabl
           noteTxtArea.setText(this.note);
      }
 
+
      public void DetailedViewSceneLoad(ActionEvent event) throws IOException {
-          String filePath = "src/main/resources/org/example/hakmana/Scene/DevDetailedView.fxml";
-          try {
-               Parent sceneRoot = FXMLLoader.load(new File(filePath).toURI().toURL());
-               stage = (Stage)((Node)event.getSource()).getScene().getWindow();
-               scene = new Scene(sceneRoot);
-               stage.setScene(scene);
-               stage.show();
-          } catch (MalformedURLException e) {
-               throw new RuntimeException(e);
-          }
+          Parent sceneroot = FXMLLoader.load(getClass().getResource("Scene/DevDetailedView.fxml"));
+          stage = (Stage)((Node)event.getSource()).getScene().getWindow();
+          scene = new Scene(sceneroot);
+          stage.setScene(scene);
+          stage.show();
      }
