+++ src/main/java/org/example/hakmana/MainApplication.java
@@ -13,7 +13,7 @@ public class MainApplication extends Application {
 
     @Override
     public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(MainApplication.class.getResource("Scene/loginPage.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(MainApplication.class.getResource("Scene/login.fxml"));
         Image icon = new Image(getClass().getResourceAsStream("Scene/Images/icon-airplay.png"));
         Scene scene = new Scene(fxmlLoader.load());
         stage.setTitle("EDMSystem");
