+++ src/main/java/org/example/hakmana/PathFinderController.java
@@ -1,4 +1,4 @@
-package org.example.hakmana.componentControllers;
+package org.example.hakmana;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -14,9 +14,7 @@ import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 
-import java.io.File;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
@@ -55,13 +53,7 @@ public class PathFinderController extends VBox implements Initializable {
 
     public PathFinderController() {
         super();
-        String filePath = "src/main/resources/org/example/hakmana/Component/PathFinder.fxml";
-        FXMLLoader fxmlPathLoader = null;
-        try {
-            fxmlPathLoader = new FXMLLoader(new File(filePath).toURI().toURL());
-        } catch (MalformedURLException e) {
-            throw new RuntimeException(e);
-        }
+        FXMLLoader fxmlPathLoader = new FXMLLoader(getClass().getResource("Component/PathFinder.fxml"));
         fxmlPathLoader.setController(this);
         fxmlPathLoader.setRoot(this);
         try{
@@ -91,16 +83,14 @@ public class PathFinderController extends VBox implements Initializable {
         continuousPressed=false;
     }
     public void goBack(ActionEvent event) throws IOException {
-        String listScenename = "src/main/resources/org/example/hakmana/Scene/dashboard.fxml";
+        String listScenename = "Scene/dashboard.fxml";
         if(!sceneList.isEmpty()) {
-            System.out.println(sceneList);
             //To remove current scene from the list.
             //Because current scene is also added to the list
             if (continuousPressed == false) {
                 sceneList.removeLast();
                 continuousPressed = true;
             }
-            //System.out.println(sceneList);
             listScenename = sceneList.getLast();
             sceneList.removeLast();
             sceneListCounter--;
@@ -109,14 +99,11 @@ public class PathFinderController extends VBox implements Initializable {
             System.out.println("list is empty");
         }
 
-        try {
-            Parent sceneRoot = FXMLLoader.load(new File(listScenename).toURI().toURL());
-            stage = (Stage)((Node)event.getSource()).getScene().getWindow();
-            scene = new Scene(sceneRoot);
-            stage.setScene(scene);
-            stage.show();
-        } catch (MalformedURLException e) {
-            throw new RuntimeException(e);
-        }
+        Parent root = FXMLLoader.load(getClass().getResource(listScenename));
+        stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
+        scene = new Scene(root);
+        stage.setScene(scene);
+        stage.show();
+
     }
 }
