--- src/main/java/org/example/hakmana/model/PhotocpyMchine.java
@@ -1,4 +1,0 @@
-package org.example.hakmana.model;
-
-public class PhotocpyMchine {
-}
