+++ src/main/resources/org/example/hakmana/Component/addDevButton.fxml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!--******************************************************************-->
-<!--              fxml file for add new device button                 -->
-<!--******************************************************************-->
-
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
