+++ src/main/resources/org/example/hakmana/Component/dashboardCard.fxml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!--******************************************************************-->
-<!--              fxml file for dashboard summary card                -->
-<!--******************************************************************-->
-
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ScrollPane?>
