--- src/main/resources/org/example/hakmana/Scene/CSS/CreateAccount.css
@@ -1,14 +1,0 @@
-/* Additional styles specific to the Create New Account interface */
-.create-account-container {
-    -fx-spacing: 10px;
-    -fx-alignment: center;
-}
-
-.create-account-field {
-    -fx-pref-width: 300px;
-    -fx-max-width: 300px;
-}
-
-.create-account-button {
-    -fx-background-color: #4CAF50;
-}
