--- src/main/resources/org/example/hakmana/Scene/CSS/EditProfile.css
@@ -1,14 +1,0 @@
-/* Additional styles specific to the Edit Your Profile interface */
-.edit-profile-container {
-    -fx-spacing: 10px;
-    -fx-alignment: center;
-}
-
-.edit-profile-field {
-    -fx-pref-width: 300px;
-    -fx-max-width: 300px;
-}
-
-.edit-profile-button {
-    -fx-background-color: #2196F3;
-}
