--- src/main/resources/org/example/hakmana/Scene/CSS/ShowUsers.css
@@ -1,16 +1,0 @@
-/* ShowUsers.css */
-@import 'UserManagement.css';
-
-/* Additional styles specific to the Show Other Users interface */
-.show-users-container {
-    -fx-spacing: 10px;
-    -fx-alignment: center;
-}
-
-.show-users-list {
-    -fx-pref-width: 300px;
-    -fx-max-width: 300px;
-    -fx-pref-height: 400px;
-    -fx-border-color: #ccc;
-    -fx-border-width: 1px;
-}
