--- src/main/resources/org/example/hakmana/Scene/DialogBoxView/CreateAccount.fxml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ButtonType?>
-<?import javafx.scene.control.DialogPane?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<DialogPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
-   <content>
-      <AnchorPane fx:id="parentAnchor" prefWidth="300.0" style="-fx-background-color: #00000021;" stylesheets="@../CSS/EditProfile.css">
-          <children>
-                      <VBox fx:id="createAccountFields" alignment="TOP_CENTER" layoutX="20.0" layoutY="20.0" prefWidth="300.0" spacing="10" styleClass="create-account-field" stylesheets="@../CSS/CreateAccount.css" AnchorPane.bottomAnchor="20.0" AnchorPane.leftAnchor="10.0" AnchorPane.rightAnchor="15.0" AnchorPane.topAnchor="0.0">
-                          <children>
-                            <!-- Insert fields and button for account creation -->
-                            <Label fx:id="createAccountLabel" text="Create New Account">
-                     <font>
-                        <Font name="System Bold" size="18.0" />
-                     </font>
-                  </Label>
-                  <VBox alignment="CENTER_LEFT" spacing="5.0" VBox.vgrow="NEVER">
-                     <children>
-                                    <TextField fx:id="newUsernameField" maxWidth="322.0" prefHeight="25.0" prefWidth="322.0" promptText="Username" styleClass="create-account-field" stylesheets="@../CSS/CreateAccount.css" />
-                        <TextField promptText="Full Name" />
-                        <TextField promptText="Post" />
-                        <TextField prefHeight="25.0" prefWidth="291.0" promptText="Employee ID" />
-                                    <PasswordField fx:id="newPasswordField" promptText="Password" styleClass="create-account-field" stylesheets="@../CSS/CreateAccount.css" />
-                                    <TextField fx:id="newEmailField" promptText="Email" styleClass="create-account-field" stylesheets="@../CSS/CreateAccount.css" />
-                        <TextField promptText="Phone Number" />
-                                    <Button fx:id="createAccountButton" styleClass="create-account-button" stylesheets="@../CSS/CreateAccount.css" text="Create Account" />
-                     </children>
-                  </VBox>
-                          </children>
-                      </VBox>
-          </children>
-      </AnchorPane>
-   </content>
-   <buttonTypes>
-      <ButtonType fx:constant="CANCEL" />
-   </buttonTypes>
-</DialogPane>
