+++ src/main/resources/org/example/hakmana/Scene/printerForm.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
 
-<DialogPane maxHeight="650.0" maxWidth="1254.0" minHeight="650.0" minWidth="1152.0" prefHeight="650.0" prefWidth="1251.0" style="-fx-background-color: #FAFAFA;" styleClass="dialogbox" stylesheets="@dialog.css" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.hakmana.view.PrinterFormController">
+<DialogPane maxHeight="650.0" maxWidth="1254.0" minHeight="650.0" minWidth="1152.0" prefHeight="650.0" prefWidth="1251.0" style="-fx-background-color: #FAFAFA;" styleClass="dialogbox" stylesheets="@dialog.css" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.hakmana.PrinterFormController">
     <expandableContent>
         <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="320.0" />
     </expandableContent>
